requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
awscli==1.19.80
awsebcli==3.19.4
bcrypt==3.2.0
blessed==1.18.0
botocore==1.20.80
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
cryptography==3.1
distlib==0.3.1
Django==3.1.2
django-cors-headers==3.7.0
djangorestframework==3.12.2
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
filelock==3.0.12
future==0.16.0
idna==2.10
jellyfish==0.8.2
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.2.2
paramiko==2.7.2
pathspec==0.5.9
pexpect==4.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4.1
requests==2.24.0
rsa==4.7.2
s3transfer==0.4.2
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
urllib3==1.25.10
virtualenv==20.0.31
wcwidth==0.1.9
websocket-client==0.59.0
